refactor: rename to projectProfile
