+++ wondrous-app/components/ProjectProfile/PodCards.tsx
@@ -1,9 +1,9 @@
 import Grid from '@mui/material/Grid';
 import Button from 'components/Button';
 import PodIcon from 'components/Icons/podIcon';
-import CreateButton from 'components/Project/CreateButton';
-import HeaderTitle from 'components/Project/HeaderTitle';
-import PodCard from 'components/Project/PodCard';
+import CreateButton from 'components/ProjectProfile/CreateButton';
+import HeaderTitle from 'components/ProjectProfile/HeaderTitle';
+import PodCard from 'components/ProjectProfile/PodCard';
 import isEmpty from 'lodash/isEmpty';
 import styled from 'styled-components';
 import palette from 'theme/palette';
